+++ src/pages/api/user/[id]/index.ts
@@ -1,5 +1,5 @@
 import crypto from 'crypto';
-import { ErrorResponse, UpdatedResponse } from '@/features/apiTypes';
+import { ErrorResponse, UpdatedResponse } from '@/common/types';
 import {
     GetUserResponse,
     deleteUserSchema,
