+++ src/pages/api/user/[id]/lock.ts
@@ -1,5 +1,5 @@
 import crypto from 'crypto';
-import { ErrorResponse, UpdatedResponse } from '@/features/apiTypes';
+import { ErrorResponse, UpdatedResponse } from '@/common/types';
 import { getUserSchema, lockUserSchema } from '@/features/user/userTypes';
 import prisma from '@/utils/prisma';
 import { mapHandlers, Handler } from '@/utils/handler';
